Extract BalanceAndActions component; use it in PortfolioScreen
